echoeats-chat/server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.54.0
langchain-openai
elevenlabs==2.20.1
python-multipart==0.0.9
